refactor(Exceptions): Renomeando o pacote corretamente.
