+++ src/main/java/com/modelsapp/models_api/Exceptions/ModelException.java
@@ -1,0 +1,11 @@
+package com.modelsapp.models_api.Exceptions;
+
+public class ModelException extends Exception {
+    public ModelException(String message) {
+        super(message);
+    }
+
+    public ModelException(String message, Exception cause) {
+        super(message, cause);
+    }
+}
