+++ src/main/java/com/modelsapp/models_api/Exceptions/RequestException.java
@@ -1,0 +1,12 @@
+package com.modelsapp.models_api.Exceptions;
+
+public class RequestException extends Throwable {
+
+    public RequestException(String message) {
+        super(message);
+    }
+
+    public RequestException(String message, Exception cause) {
+        super(message, cause);
+    }
+}
