+++ src/main/java/com/modelsapp/models_api/Exceptions/UserException.java
@@ -1,11 +1,11 @@
-package com.modelsapp.models_api.Execptions;
+package com.modelsapp.models_api.Exceptions;
 
 public class UserException extends RuntimeException {
     public UserException(String message) {
         super(message);
     }
 
-    public UserException(String message, Throwable cause) {
+    public UserException(String message, Exception cause) {
         super(message, cause);
     }
 }
